SRTP-769-implement-api-send-gpd-message
