--- gpd-test/force-release
@@ -1,1 +1,0 @@
-0000
