+++ gpd-test/main.py
@@ -1,22 +1,18 @@
-from fastapi import FastAPI, HTTPException, Request
+from fastapi import FastAPI
 from contextlib import asynccontextmanager
-from producer import setup_producer
-from dto import RTPMessage
 from dotenv import load_dotenv
-import os, json
+from producer import setup_producer
+from gpdApi import gpd_router
+from healthCheckApi import health_router
 import logging
-
-# Setup logging
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
 
 load_dotenv()
 
-EVENTHUB_TOPIC = os.environ.get("EVENTHUB_TOPIC", "default-topic")
-
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
 
 @asynccontextmanager
-async def app_lifespan(fastapi_app: FastAPI):
+async def app_lifespan(app: FastAPI):
     try:
         logger.info("Starting producer setup...")
         kafka_producer = await setup_producer()
@@ -28,99 +24,9 @@ async def app_lifespan(fastapi_app: FastAPI):
         logger.info("Producer shutdown completed")
     except Exception as e:
         logger.error(f"Error during lifespan: {e}")
-        # Set a flag to indicate producer is not available
         app.state.producer = None
         yield
 
-
 app = FastAPI(lifespan=app_lifespan)
-
-def sanitize_log_value(value) -> str:
-    """Sanitize a log field by removing newlines and carriage returns."""
-    return str(value).replace("\n", "").replace("\r", "")
-
-
-@app.get("/health")
-async def health_check():
-    """Health check endpoint for Kubernetes probes"""
-    return {"status": "healthy", "service": "gpd-producer"}
-
-
-
-# @app.post("/send/gpd/message")
-# async def send_msg(message: RTPMessage, request: Request):
-#     logger.info(
-#         "Received message with id=%s, operation=%s, status=%s",
-#         sanitize_log_value(message.id),
-#         sanitize_log_value(message.operation),
-#         sanitize_log_value(message.status),
-#     )
-#
-#     try:
-#         producer = request.app.state.producer
-#         if producer is None:
-#             raise ConnectionError("Kafka producer is not available")
-#
-#         await producer.send_and_wait(
-#             EVENTHUB_TOPIC,
-#             json.dumps(message.model_dump(by_alias=True)).encode("utf-8")
-#         )
-#
-#         logger.info(f"Message sent successfully")
-#         return {
-#             "status": "success"
-#         }
-#
-#     except Exception as e:
-#         logger.exception("Error sending message: %s", e)
-#         raise HTTPException(
-#             status_code=500,
-#             detail={
-#                 "status": "error",
-#                 "message": str(e)
-#             }
-#         )
-
-from fastapi import Request, Query
-
-@app.post("/send/gpd/message")
-async def send_msg(request: Request, validate: bool = Query(default=True)):
-    payload = await request.json()
-
-    logger.info("Received message: %s", sanitize_log_value(payload))
-
-    if validate:
-        try:
-            _ = RTPMessage(**payload)  # just to validate
-        except Exception as e:
-            logger.warning("Validation failed: %s", e)
-            raise HTTPException(
-                status_code=422,
-                detail={
-                    "status": "error",
-                    "message": f"Payload validation failed: {e}"
-                }
-            )
-
-    try:
-        producer = request.app.state.producer
-        if producer is None:
-            raise ConnectionError("Kafka producer is not available")
-
-        await producer.send_and_wait(
-            EVENTHUB_TOPIC,
-            json.dumps(payload).encode("utf-8")
-        )
-
-        logger.info("Message sent successfully")
-        return { "status": "success" }
-
-    except Exception as e:
-        logger.exception("Error sending message: %s", e)
-        raise HTTPException(
-            status_code=500,
-            detail={
-                "status": "error",
-                "message": str(e)
-            }
-        )
+app.include_router(health_router)
+app.include_router(gpd_router)
